auto: fix perl errors
The error ". is no longer in @INC" in Perl occurs because, as of Perl version 5.26, the current directory is no longer automatically included in @INC, the list of directories where Perl looks for modules.

This change was implemented to improve security. The fix is to explicitly tell Perl where to find the local script.